# docs-final/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a customer or service-provider account and immediately sends a one-time verification code (OTP) to the phone number you provided. Finish signing up by calling POST /auth/otp/verify with purpose \"REGISTER\" and that code.\n\n**Testing locally:** since no real SMS provider is configured, read the code back from `GET /dev/inbox` instead of a real phone.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "CUSTOMER",
                      "PROVIDER"
                    ]
                  },
                  "phoneE164": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{7,14}$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  },
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "lastName": {
                    "type": "string",
                    "maxLength": 80,
                    "default": ""
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ur"
                    ],
                    "default": "en"
                  }
                },
                "required": [
                  "role",
                  "phoneE164",
                  "password",
                  "firstName"
                ],
                "additionalProperties": false
              },
              "examples": {
                "customer": {
                  "summary": "Customer",
                  "value": {
                    "role": "CUSTOMER",
                    "phoneE164": "+923001234567",
                    "password": "CorrectHorse9Battery",
                    "firstName": "Ayesha",
                    "lastName": "Khan"
                  }
                },
                "provider": {
                  "summary": "Service provider",
                  "value": {
                    "role": "PROVIDER",
                    "phoneE164": "+923001234568",
                    "password": "CorrectHorse9Battery",
                    "firstName": "Bilal",
                    "lastName": "Ahmed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Sign up for a new account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "description": "Sends a short-lived numeric code by SMS or email to the given phone number or email address. Used to verify a new account, to log in without a password, or to start a password reset — the \"purpose\" field tells the server which of these you are doing.",
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "target": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 320
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "REGISTER",
                      "LOGIN",
                      "PASSWORD_RESET",
                      "PHONE_CHANGE"
                    ]
                  }
                },
                "required": [
                  "target",
                  "purpose"
                ],
                "additionalProperties": false
              },
              "examples": {
                "login": {
                  "summary": "Request a login code",
                  "value": {
                    "target": "+923001234567",
                    "purpose": "LOGIN"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Send a one-time verification code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "description": "Submits the code you received by SMS or email. For the REGISTER and LOGIN purposes, a correct code logs you in and returns an access token; for other purposes it just confirms the code matches.\n\n**Testing locally:** find the 6-digit code in `GET /dev/inbox`.",
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "target": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 320
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "REGISTER",
                      "LOGIN",
                      "PASSWORD_RESET",
                      "PHONE_CHANGE"
                    ]
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4,8}$"
                  }
                },
                "required": [
                  "target",
                  "purpose",
                  "code"
                ],
                "additionalProperties": false
              },
              "examples": {
                "register": {
                  "summary": "Confirm registration",
                  "value": {
                    "target": "+923001234567",
                    "purpose": "REGISTER",
                    "code": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Verify a one-time code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Signs in using a phone number or email plus a password. Staff accounts (admin, finance, agent) must also supply a valid two-factor (TOTP) code — see the 401 response below.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "totpCode": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "identifier",
                  "password"
                ],
                "additionalProperties": false
              },
              "examples": {
                "customer": {
                  "summary": "Customer or provider",
                  "value": {
                    "identifier": "+923001234567",
                    "password": "CorrectHorse9Battery"
                  }
                },
                "staff": {
                  "summary": "Seeded admin (dev)",
                  "value": {
                    "identifier": "admin@smart-home.local",
                    "password": "DevPassword!2026",
                    "totpCode": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Either the email/phone and password did not match, or (for staff accounts) a two-factor authentication code is required to finish logging in."
          }
        },
        "summary": "Log in with a password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Uses the httpOnly refresh cookie set at login to issue a fresh access token without asking the user to log in again. Each refresh token can only be used once, and using it issues a new one.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 200
                  }
                },
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "Non-browser client holding the token itself",
                  "value": {
                    "refreshToken": "a-refresh-token-from-a-previous-login"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "The refresh token had already been used once before. That looks like the token was stolen and replayed, so every session descended from it has been signed out as a precaution — the user needs to log in again."
          }
        },
        "summary": "Get a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Ends the current session by invalidating the refresh cookie. The access token you were holding will simply expire on its own shortly after (it is not individually revoked).",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 200
                  }
                },
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "Non-browser client holding the token itself",
                  "value": {
                    "refreshToken": "a-refresh-token-from-a-previous-login"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Log out",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "description": "Sends a reset code to the given phone number or email if an account exists for it. To stop attackers from being able to guess which accounts exist, this endpoint always returns the same response whether or not a matching account was found.",
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 320
                  }
                },
                "required": [
                  "identifier"
                ],
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "By phone",
                  "value": {
                    "identifier": "+923001234567"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Request a password reset code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "description": "Submits the reset code from POST /auth/password/forgot along with a new password. On success, every other active session on the account is signed out for safety, and this response logs you in with a fresh session.",
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 320
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4,8}$"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  }
                },
                "required": [
                  "identifier",
                  "code",
                  "newPassword"
                ],
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "Reset with the emailed/texted code",
                  "value": {
                    "identifier": "+923001234567",
                    "code": "123456",
                    "newPassword": "BrandNewPass9"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reset your password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Returns the currently logged-in user's details, including their roles (e.g. CUSTOMER, PROVIDER, ADMIN) and, for providers, their approval status. Useful right after login to know what the user is allowed to do.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get your own profile",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/totp/setup": {
      "post": {
        "description": "Generates a new secret key and QR-code link for an authenticator app (e.g. Google Authenticator, Authy). This is shown only this one time, so save it — then confirm it works with POST /auth/totp/verify before it is treated as active.",
        "operationId": "AuthController_totpSetup",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start setting up two-factor authentication",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/totp/verify": {
      "post": {
        "description": "Verifies the 6-digit code currently showing in your authenticator app. If it matches, two-factor authentication is switched on for the account from now on.",
        "operationId": "AuthController_totpVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "code"
                ],
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "Authenticator code",
                  "value": {
                    "code": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Confirm and turn on two-factor authentication",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/totp": {
      "delete": {
        "description": "Disables two-factor authentication for the account. Sessions that are already logged in keep working, but will be asked to set TOTP up again the next time they log in as staff.",
        "operationId": "AuthController_totpDisable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Turn off two-factor authentication",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Home Maintenance Services API",
    "description": "The backend API for the Smart Home Maintenance Service platform: account sign-up and login, admin settings, payment webhooks, and health checks.\n\n**Getting started:** most endpoints need you to be logged in. Call `POST /auth/login` (password) or the OTP endpoints to get an `accessToken`, then click \"Authorize\" above and enter it as `Bearer <token>` to unlock the endpoints marked with a lock icon.\n\nEndpoints under **development** only work when the API is running with `DEV_INBOX_ENABLED=true` (local/dev setups) — they let you see the text messages, emails and files the mock providers would otherwise send to real services, which is how you retrieve OTP codes while testing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}